requirements.txt
requests>=2.31.0
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
